added backgrounds and started hearts & more assets

+++ Assets/Scripts/Player/Combat.cs
@@ -3,8 +3,7 @@ using System.Collections;
 
 public class Combat : MonoBehaviour
 {
-    [SerializeField] private int attackDamage = 10;
-
+    [SerializeField] private float attackDamage = 10f;
     private Animator anim;
     private bool isAttacking = false;
